feat: 책 상태관리 분리, 상태관리 post api연동
